Added authorization to routes

+++ go_rest_api/models/event.go
@@ -12,7 +12,7 @@ type Event struct {
 	Description string    `binding:"required"`
 	Location    string    `binding:"required"`
 	DateTime    time.Time `binding:"required"`
-	UserID      int
+	UserID      int64
 }
 
 func (e *Event) Save() error {
@@ -20,6 +20,7 @@ func (e *Event) Save() error {
 	INSERT INTO events(name, description, location, dateTime, user_id)
 	VALUES (?, ?, ?, ?, ?)
 	`
+
 	stmt, err := db.DB.Prepare(query)
 
 	if err != nil {
